requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
assemblyai==0.43.1
certifi==2025.8.3
click==8.2.2
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
